Equipamento: Move service e classe para seu diretório ideal
