+++ src/app/classes/equipamento.spec.ts
