+++ src/app/classes/equipamento.ts
