+++ src/app/services/equipamento.service.spec.ts
